Fix for Windows print related decoding error
Summary:
Fixes encoding error introduced by printing certain
characters to the terminal in the next diff on Windows
machines by writing directly to a file rather than printing
to the console. This fix is needed because it is not possible to
print all utf-8 characters to the console on windows.

Reviewed By: avp

Differential Revision: D29852064

fbshipit-source-id: 9d44b7d6a820d0c2988ee1450c1c7cdfd377be26