Pipe constructor in pipe_wrap.cpp
Summary:
Added the pipe constructor to pipe_wrap.cpp.
This, along with the newly commented in code
in net.js, should give js the ability construct the
Pipe object that holds the uv_pipe_t stream
handle.

Made a class inside of pipe_wrap.cpp the same
way we did for tty_wrap.cpp to create a HostObject
to store the handle. Updated RuntimeState to save
this object's name as a PropNameID.

Reviewed By: avp

Differential Revision: D29746437

fbshipit-source-id: 68d07d2d9cf2c3fbecf0393d3e2d269f7f43abe7