Use local types directory instead of @babel/types.
Summary:
Include a partial set of the files from `babel/types` locally.
These can be used directly instead of `babel/types` throughout
the hermes-parser package.

Export a partial set of the babel/types API which can be used
to handle the existing functionality in transformFromAstSync.

Some parts of this are currently just stubs.
For example, `validate` is just an empty function for now,
as the goal is just to make the existing tests run with types
generated by the Hermes parser/AST instead of Babel's version.

Some minor functions needed to be manually implemented or
adapted (e.g. `isType`).

Reviewed By: pieterv

Differential Revision: D28802064

fbshipit-source-id: 7c486487b717a8a29c8f8b8bedf21e01a4291d42