Add babel-traverse dependency.
Summary:
Add the `babel/traverse` package as a dependency based on a Flow-based
commit of the package.
Include the original license in `BABEL_LICENSE`,
and add a `README.md` which explains where the code came from.
Add the headers to allow Flow parsing/formatting.

This requires some minor changes in the `.babelrc` for `hermes-parser`
to ensure that the Flow types are stripped before class props
are transformed, as well as `allowDeclareFields` to enable use
of features in the Flow code in `traverse`.

The existing code includes some `"todo"` code which Babel never implemented.

Reviewed By: pieterv

Differential Revision: D28429412

fbshipit-source-id: 690b85a6b5e1e8feb7eb8975124350ca4f30db85