Clean up organisation of object literal lowering
Summary:
Collapse some of the nested expressions so that we can essentially get
four independent blocks:
1. Computed properties
2. Special proto properties
3. Accessors
4. Normal values

Move the block for proto properties above accessors, since like
computed properties, it does not have a valid value for `propValue` (so
it's better to just not have it available). This then lets us add an
assert that anything with a valid `propValue` (i.e. accessors and
normal values) must at least generate a placeholder for its property.
This should help avoid us missing cases that we need to maintain
ordering.

Reviewed By: tmikov

Differential Revision: D29222537

fbshipit-source-id: 17c5024abc83c03759d4971792e5e075deb501e0