Remove use of CompressedPointer::getLoc in WeakRoots
Summary:
We can avoid using `getLoc` by instead dealing directly in raw
pointers. This will likely be slightly less efficient when compressed
pointers are disabled, but this is not currently a performance critical
path (since it is unused in Hades). To improve efficiency in other GCs,
we would have to implement custom acceptors like Hades does.

Reviewed By: kodafb

Differential Revision: D29956058

fbshipit-source-id: 5beb458a3e64e641bffbddb49887eceabc806427