Console time functionality and test case
Summary:
Added the process.hrtime() function that can do one of
two things: if it is passed in an argument, it will output
the difference between the argument and the current
time in a [seconds, remaining nanoseconds] format.
Otherwise, it will just output the current time representation
in the [seconds, remaining nanoseconds] format.

Other changes include commenting out the trace functions.

Reviewed By: avp

Differential Revision: D30108579

fbshipit-source-id: 34b71d352ec0330c3e69ce1890596af242f069c3