Console trace functionality and test case
Summary:
Outputs the stack trace depending on where console.trace
is being called. Prints the error to stderr. Added code to
constructor.js to recreate the functionality of Node's
ErrorCaptureStackTrace. This code just creates
a new error with the correct message and also
removes the top most level of the trace (which will
be the trace function itself).

Reviewed By: avp

Differential Revision: D30108583

fbshipit-source-id: 8161db6f5bc3b1f66ea6ffe1cae9f2971f07c890