Begin AST validation logic
Summary:
Set up some new syntax in the NodeKind macro to allow specifying
parents of Node kinds, as well as constraints on each of their children.

A new `validate` function handles validating a Node and its children.
It returns a `bool` to indicate the failure to validate.
This can be used in conjunction with a visitor to recursively validate
an entire AST.

Reviewed By: tmikov

Differential Revision: D30081493

fbshipit-source-id: d9a5c6beb14b07589e5977a96e3aef3baf582e09