Simplify build for hermesapi
Summary:
Make the following simplifications:
1. Remove the `CREATE_SHARED_LIBRARY` flag. It is currently being
passed in only for the static library `hermesapi` and not for
`libhermes` (which seems incorrect). For consistency across MSVC and
non-MSVC, we might as well remove the flag and always define
`HERMES_EXPORT`.
2. Remove custom compile flags from `hermesapi`, since `libhermes`
uses the same sources and currently builds fine without them.
3. Remove explicitly setting the `CXX_STANDARD`, which is now 14
across our code.

Reviewed By: kodafb

Differential Revision: D30242173

fbshipit-source-id: 06fee12b89c74b3085a8899cfe325f78102de785