Add input source map support
Summary:
Extend `hparser` with an optional converter trait invoked for every
location. The converter can use the existing SourceMap module.

Add an integration test doing just that.

The converter is passed to the parser as a `dyn trait` (oh, no!). This
was a careful decision. These alternatives were considered and
rejected:
- Generics (either a trait or a closure). That would cause the entire
  module to be generic. Seems excessive.
- Typed erased closure. Needs a `Box`. Doesn't seem preferable.

Reviewed By: avp

Differential Revision: D30444132

fbshipit-source-id: f2ce1a5e396987fc1b3186c76a32838e43a7e6c5