Remove lastValue_ in serialised literal parser
Summary:
Caching the value of constants like null, true, and false does not
actually provide a performance benefit, since these are compile time
constants. Instead, we can significantly reduce the number of branches
in this code by "re-creating" the value on each iteration, and just
returning directly from each entry in the switch.

Reviewed By: kodafb

Differential Revision: D30523831

fbshipit-source-id: dc7e7bcba23f1c857a360f86d0984141771cd5a1