Put metadata table in global array
Summary:
Now that metadata has a constexpr default constructor and no
destructor, we can simply store it in a global array.

It can be more efficient to access the metadata table through a global
array as opposed to going through an indirection through a member of
GC. Specifically, this can be optimised better in the marking loop,
since the location of the metadata array doesn't have to be loaded
from the GC on each iteration.

Reviewed By: kodafb

Differential Revision: D28376173

fbshipit-source-id: 003026d0d3d1e1480ff9cd2e062317093e2f8db8