Delete SourceMap, Move Hermes FFI, Support into crates
Summary:
The Hermes SourceMap interface is no longer needed, so it has been
deleted.

The Hermes crate wraps all Hermes dependencies but has no dependency on
anything else Juno-related. Although such separation is not strictly
necessary, it cleans up the design.

The conversion of the Hermes AST to Juno AST remains in the Juno crate.

Support is also moved to a separate crate so it can be accessed by the
Hermes FFI. This required moving `convert` into `support`, which is
exactly the type of dependency we want to catch and prevent.

Added a workspace declaration to Cargo.toml to enable all tests to run
with `cargo test --workspace`.

Reviewed By: avp

Differential Revision: D30467585

fbshipit-source-id: f6194453e733cc9f0639e6ff1a2b3b7a1a5257f5