Replace StringCycleChecker with scope guard
Summary:
The `StringCycleChecker` is a little weird to use because it cannot
return an error from its constructor. Instead, we have to check the
result of the constructor using `foundCycle`, and have to ensure that
the stack in the runtime is correctly set up so that the destructor can
run safely. This is also inefficient, but that probably doesn't matter.

Given that there are only two places it is currently used, a scope
guard seems simpler and clearer.

Reviewed By: kodafb

Differential Revision: D29954709

fbshipit-source-id: 0eb814f4568f315453f7b0d3c63b363802157d7b