Handle use of non-serialisable literals in object literal
Summary:
Not all `Literal` types are serialisable (only 4 out of the 7 are). It
is safer to enumerate the types that can be serialised, than to check
by omitting types.

For instance, the global object is treated as a literal, but cannot be
serialised, however, since we only check for undefined, we end up
trying to serialise it and crashing or generating bad code.

Reviewed By: avp

Differential Revision: D30626995

fbshipit-source-id: 2631467b846434506f06899ef5aedea845025822