Add a root node to profiler traces
Summary:
Currently, the sampling profiler does not correctly report gaps where
the runtime is not executing any code. These samples get dropped
entirely, and as a result, tracery just extends the preceding sample to
take up the gap. This is very misleading, since it makes it appear as
though a function is taking up more time than it actually is.

To fix this, add a new root node that exists for every sample, even
those where the runtime is inactive. Now, points where the runtime is
inactive will just show up as "[root]".

Reviewed By: kodafb

Differential Revision: D30895444

fbshipit-source-id: 9240de33a24febeae70ecc39c54c432489b8317e