Remove special string printing under ICU
Summary:
This code provides special string encoding functionality, that is only
used for developer facing things (as far as I can tell) since it prints
to an `llvh::ostream`. This code was originally kept because only Apple
platforms were unable to use ICU, and we wanted the more accurate
encoding on all other platforms. However, we now only appear to do this
system specific encoding on Linux, since ICU is not included in Hermes
on Apple and Android, and the Windows version is hard-coded to UTF-8.

The motivation is that this is the only use of ICU outside
`PlatformUnicodeICU.cpp`, so removing it makes our unicode
implementation self contained.

Reviewed By: tmikov

Differential Revision: D30634531

fbshipit-source-id: b475a8173fc2947ca4c43c1b8bd61707fc699e0c