hermes/utf: export a couple of more functions
Summary:
Export a couple of minor UTF-8 related function from the hermes/utf
module. Technically they may better belong in the support crate because
they are not Hermes-specific, but I didn't want to create a whole new
module for two trivial functions, plus the customer is a customer of the
hermes_parser module as well.

Reviewed By: avp

Differential Revision: D31044168

fbshipit-source-id: 099080ba2869c9df65e8f6713878b2895f950373