Track output location during JS generation
Summary:
Add a `position` tracker field to `GenJS` which is updated whenever
we output to the writer.

We need to know how many characters we output each time.
In the majority of cases, we're just outputting ASCII,
so just check the number of bytes in the `write_ascii` case
(used by `out!`).

For now, `gen_js::generate` returns an empty `SourceMap`.
That will be made to work when we begin emitting segments for
tokens as we emit them in the source.

Reviewed By: tmikov

Differential Revision: D30546171

fbshipit-source-id: 5e584f19e5cb73f177b78f399a067953c06a19af