Merge Node and NodeKind
Summary:
Avoid the awkward distinction between Node and NodeKind
by embedding the shared information within each of the individual
structs.

In order to ensure that we have efficient access to these shared
fields, we provide getter functions to retrieve references.
These will pattern match on all the nodes and retrieve the field,
which is generated to be the first field.
Forcing `repr(C)` on all relevant structures ensures the layout
is consistent and that the identical match branches will be optimized
away.

Reviewed By: tmikov

Differential Revision: D31123068

fbshipit-source-id: 665430335f014d08c1ca3ec9464a19144d951cc5