Add an out_token! macro for emitting source mappings
Summary:
Instead of case-by-case calls to `add_segment` prior to `out!()` calls,
introduce an `out_token!()` macro that will add the segment prior
to calling `out!()`.

This makes it easier to keep track of when location info is being
output at the start of the generation for an AST node.

Reviewed By: tmikov

Differential Revision: D31254318

fbshipit-source-id: b9f8d0bd840b82b8ea3c2d3a6786df3995823b65