Integrate SourceManager in Context
Summary:
Add SourceManager as a field in Context. For now it is default
initialized, but in the future we may add a move-constructor to pass in
an existing instance.

Change `Context::add_atom()` to take immutable self and rename it to
`Context::atom()` to reflect that (it maps a string to an atom).

Refactor existing parsing to integrate with the source manager.

Reviewed By: avp

Differential Revision: D31325011

fbshipit-source-id: 63086c5c502c1d40361693c5787e2f8621be096e