add traversal tooling - SafeEmitter
Summary:
Build out the new transformation tooling.
This diff:
Forks ESLint's "SafeEmitter" - flowify it, and remove the cruft that we don't need.
This simple class is just an event emitter typed to call listeners with nodes.
The traversal logic uses this as a store where the event names are esquery selectors, and it "emits" events against selectors as it traverses the AST.

Reviewed By: pieterv

Differential Revision: D31877350

fbshipit-source-id: b13570e5e5af4449d011875fc0f80dcbcad34fdc