add traversal tooling - traverse
Summary:
Build out the new transformation tooling.
This diff:
Adds a function which consumes all of the previous code to allow you to use esquery selectors to visit the AST.
It also provides a "context" object which can be used to get some additional information during the traversal.
Currently that is just scope analysis information.

Reviewed By: pieterv

Differential Revision: D31877405

fbshipit-source-id: 6e0cf0d85ded33e12191a5e93cf0a74d6a6b11e3