add tests to ensure that the declared flow types are correct
Summary:
By leveraging the JSON that we now output - we can do a sanity check of the declared flow types.
This allows us to have some assurance that the flow types are kept in sync with the parser declarations.

This diff also fixes up the gaps in the types that were exposed by the test!

Reviewed By: pieterv

Differential Revision: D31973335

fbshipit-source-id: bc623b43495c8bd482bee300eb621ee9dd50b41f