add codegen step to create a JSON blob for estree defs
Summary:
Doing complex JS codegen using the C++ preprocessor is very cumbersome:
- we have to define 0-8 macros, often duplicating lines
- C++ preprocessor macros are filled with various pitfalls
- we can't do conditional generation based on input
- you have to pass in the C++ include paths to run the js script

Considering our scripts are written in JS - we don't really need a level of indirection. We just need to get the raw data into JS, then we can build our files using the full power of the language.

This diff:
- adds a generation step which generates a JSON blob from the ESTree defs.
- adds `babel-node` so that we can write flow-typesd scripts

In the next diff I'll start converting the existing codegen steps to be pure JS scripts AND to be driven from this blob

Reviewed By: pieterv

Differential Revision: D31940978

fbshipit-source-id: 70d792e7b185c404c20bfd709f72c54ad8e51c1a