Treat SymbolID barrier as a snapshot barrier
Summary:
`writeBarrier(SymbolID)` looks and behaves exactly like the other
snapshot barriers, since it is only interested in the old value, so its
name is somewhat misleading. Change its name and interface to bring it
in line with the other snapshot write barriers.

This change also inlines the fast path of the write barrier, and adds
the `inYoungGen` check to elide the write barrier, which makes the code
consistent and could be a small perf bump.

Reviewed By: kodafb

Differential Revision: D31879541

fbshipit-source-id: 4df4ae502047d40082ee3fa2f63e1fa2a507fad9