Garbage collected AST
Summary:
The AST now is garbage collected and managed by `Context`.
Nodes point to other nodes with references (`&Node<'a>`).
`Node` cannot be constructed directly by users,
instead they must use various `Builder` structs and pass them
`Template` structs to initialize new nodes,
which will then be allocated on the GC heap in `Context`.
Nodes themselves are immutable once placed on the GC heap.

Use of the actual `Node`s is facilitated by `GCContext`,
which provides `&'gc Node<'gc>` where `'gc` is the lifetime of the
`GCContext` itself.

Mutating and non-mutating visitors are provided.
Mutating visitor simply returns a new `Node` which is the mutated
AST, and the old AST nodes will be collected on the next GC cycle.

See the top-level comment in ast/mod.rs for more information.

Reviewed By: tmikov

Differential Revision: D31563998

fbshipit-source-id: df6849f6983b17e4d1f10fc83b0f0d90e08b5d92