Make target size a moving average
Summary:
Dynamically adjusting the collection threshold means that we operate
much closer to the target size, since we can start collections fairly
late. Since the target size is currently strictly a function of the
amount of live data at the end of the previous collection, it can
fluctuate by a few MB each time it is updated.

Trying to track that number exactly adds extra work, since we need to
constantly compact away segments and then add them back. Instead, we
can use a moving average to smooth out the target size, so that small
fluctuations do not trigger immediate action from the GC.

Reviewed By: kodafb

Differential Revision: D31987201

fbshipit-source-id: d0b1b77ee6781545700ad940ca44512f53882bcf