Teach the resolver about well known globals
Summary:
If `warn_undefined` is enabled, pre-define a list of well known common
global properties, to decrease the number of generated warnings.

This is not expected to be used in production, but is useful for
testing. Also note that it does not possibly affect correctness in any
way, it only affects whether an ignorable warning is displayed. So there
is no harm in adding more "well known" globals.

Reviewed By: avp

Differential Revision: D32056903

fbshipit-source-id: a326ca1096b195f6617d992c402085b73964cfe2