SemContext: all semantic resolution data for an AST
Summary:
SemContext contains the semantic resolution data for an AST: lexical
scopes, declarations, mapping from identifiers to declarations
(bindings), etc.

This is not intended to be the final form of this API, because it is not
yet fully understood how it will be used. Also, this form of index based
API is relatively unfamiliar.

Reviewed By: avp

Differential Revision: D32040068

fbshipit-source-id: 3a74fea81ea1de269244919f17ae19747aa5f686