a library for CLI parsing.
Summary:
`command_line` is inspired by the API `llvm::CommandLine`, although it
doesn't follow it closely and the implementations are completely
different.

It is meant to be a very simple library for handling most common tasks,
where simple things work naturally and by default.

It does no tricks and doesn't worry much about memory or performance,
since those hardly matter in command line handling. Unsafe code is used
in only one place in a clearly defined way (for UnsafeCell).

It is intended to replace Structopt in Juno, since Structopt makes some
natural use cases almost impossible, or at least very difficult.

`command_line` works by constructing `Opt<T>` objects, which serve as
holders of the command line argument. This is similar to
`llvm::CommandLine`. A more detailed documentation will be added in
future diffs. Since it is written in Rust, it can rely on
`..Default::default()` to skip default fields, instead of the template
tricks that LLVM is forced to use.

Note that a wrapper similar to Structopt can be written on top of it,
but there probably wouldn't be much point. Even if it looks verbose,
options are not added to a program every day.

Reviewed By: avp

Differential Revision: D32242090

fbshipit-source-id: 470bd1464a80961e1edc7ddbaf9bce4074e70a8c