Use a lookup table for return opcode
Summary:
Instead of generating a series of branches and relying on the compiler
to do something efficient, we can eliminate the branches entirely by
using a lookup table packed into a `uint32_t`. While this is slightly
more expensive in terms of instructions, it produces no branches and
results in a net perf win.

Reviewed By: tmikov

Differential Revision: D32229103

fbshipit-source-id: 9a332e51d0d910c41e4ee07e3146efbc61618a12