fork prettier comment attachment codebase
Summary:
This diff simply forks prettier's comment attachment code.
We want to use their algorithms, but it's not an exposed API.

Pieter has a PR for prettier which exposes it within the formatting custom parser - but if we consume it that way then we'd be stuck always prettier over every file, instead of just on the files with mutations - which is a big waste of time during large transforms.

Reviewed By: pieterv

Differential Revision: D32263740

fbshipit-source-id: 2a3b990978a12bcaf71adc6b7a5e9434c08787c9