Add replace_with_multiple
Summary:
Add a function to replace a single node with multiple by using a new
`TransformResult::Expanded` variant.
Replacing with multiple will panic if the parent is not a `NodeList`.
It will soon be easier to check this via knowing where the current
node resides in the parent.

Update the type stripping pass to use this.

Reviewed By: tmikov

Differential Revision: D32330693

fbshipit-source-id: 23ec532622f628be8023cf6e2010ebe380ec5518