fix codegen based on usage testing
Summary:
first - when writing utils it's not uncommon to spread two sets of query selectors together.
due to the indexer - flow will error on such spreads. this diff just exposes a version with and without the selector.

second - the selector was previously typed as `:ESNode`. The problem with this is that if you have a selector that we can't statically type - we want to manually type it like `Foo[id="bar"]`(node: Foo)`. However this "refinement" isn't sound - so flow blocks it. This makes the types awful to use! Thus instead of making it a sound type, we type it as `any` and trust the user to properly type it.

third - I noticed the codegen was wrong! babel wasn't translating generics properly. It turns out that without the `flow` - babel will see `foo<T>()` as `foo < T > ()` (eg binary expressions!). So all of the codegen now includes a loose flow by default so that babel will correctly transform it.

Reviewed By: evanyeung

Differential Revision: D32475889

fbshipit-source-id: 464ab89493e8efaaebc60f728b0e4efd8d42f214