add codegen for strict shallowCloneNode typing
Summary:
One thing I noticed when consuming the shallow clone API was that it was typed wrong.
I typed it early on in the design before we'd introduced detached nodes.

Unfortunately dynamically typing won't work here - but thankfully we've already got the types codegenned!
But unfortunately there's no way to dynamically lookup the types :(
So more codegen - but this time codegen based on the other codegen!

Reviewed By: pieterv

Differential Revision: D32544135

fbshipit-source-id: 0a8432f9160a760c9be600c6e0f737c6c246cf49