Document that WeakRefs only currently support pointers
Summary:
The ability of WeakRefs to store non-pointer values is currently pretty
unclear. While the functions accept any `HermesValue`, until recently,
the tagged pointer implementation of `WeakRefSlot` meant that it wasn't
possible to actually store a non-pointer. In addition, the read barrier
doesn't handle symbols properly.

For now, we should at least assert that the value must be a pointer. If
we want to extend it to other things, the read barrier should be fixed.

Reviewed By: avp

Differential Revision: D32873460

fbshipit-source-id: f9d5af2fbc76ebc0c4c29dd89c46e4b9ebbfec1a