fix codegen and add some basic codegen validation
Summary:
I found a bug in the codegen where an undefined type name was included when it shouldn't have been.
For whatever reason flow didn't complain about that at the consumption site - and instead just treated it as `any` - which caused confusing errors.

This diff fixes that bug and adds a basic validation step to the codegen using ESLint to check for:
- unused vars (potentially indicating missing codegen)
- usage of undefined vars (definitely indicating missing codegen)
- var redeclaration (potentially indicating duplicate codegen)

Reviewed By: pieterv

Differential Revision: D33535255

fbshipit-source-id: e08bd4dcba9e47e0b707b94b6446d7962fabbff7