add mutation for removing a node (with constrained types)
Summary:
This is the first cut of a more general "remove node" API.
We originally didn't include this because it's not easy to provide a "general-purpose" API - there are many different cases and caveats to guard against.

This initial API allows you to remove the following "classes" of nodes:
- Class members
- Function params
- Object expression properties
- Array expression elements
- Object type properties (which includes interface bodies)
- Enum members

These are all easy to handle because their "container" is an array of items on the parent where it is safe to have an empty array.

In future we could consider handling more cases if required.

Reviewed By: pieterv

Differential Revision: D33539844

fbshipit-source-id: c447a48b8fcbeb039b88350907a9047a2dda5eb7