Fix id tracker access from background thread
Summary:
`objectIDMap_` must only be accessed when `mtx_` is held.

Note that this is slightly different from checking whether the
AllocationLocationTracker or SamplingAllocationTracker are enabled. In
the latter cases, we explicitly guarantee that the background thread is
not running when the enabled flag is modified. However, `objectIDMap_`
may be manipulated at any time.

Reviewed By: mhorowitz

Differential Revision: D33838429

fbshipit-source-id: 8932550a4ed9ea3b18fc7cf5d8052dd3570c4eb3