update estree literal representation so it's easier to work with
Summary:
Unfortunately flow doesn't keep `typeof` refinements through `%checks`.
So attempting to use flow to refine a `Literal` to one of the specific types is impossible.

This change adds a new `literalType` property to the nodes which allow us to easily refine the type thanks to the discriminated union.
This change also fixes the `BigInt` node representation so it matches the spec by emitting the raw bigint value if supported by the runtime.

Reviewed By: pieterv

Differential Revision: D33931798

fbshipit-source-id: 0dc3f2d53a926e776d033811917b4da5db00c02a