Implement Error.captureStackTrace
Summary:
Implements V8's [`Error.captureStackTrace`](https://v8.dev/docs/stack-trace-api#stack-trace-collection-for-custom-exceptions) API in Hermes.

## `Error.captureStackTrace(target, functionOpt)`

Collects a stack trace and adds a `stack` getter to the given `target` object that returns that stack trace, using an identical format to that of the `stack` property of an `Error` instance.

If the second argument to `captureStackTrace` is a function, when collecting the stack trace all frames above the topmost call to the provided function, including that call, are left out of the stack trace.

NOTE: The V8 docs state that `captureStackTrace` formats traces eagerly when called, but its actual behaviour is to install a lazy `stack` getter, like we do here.

## Implementation details

When `captureStackTrace` is called, it:
1. Creates a bare-bones `Error` instance holding the stack trace and attaches it to `target` under the named internal property `[[CapturedError]]`.
2. If `functionOpt` is given and is callable:
Skips frames by searching for the code block associated with `functionOpt` - its own if it's a `JSFunction`, or that of its target if it's a `BoundFunction` (and so on).
3. Defines a `stack` property on `target` using the exact same accessor that `Error` instances use for this purpose.
The getter now distinguishes between the *error object* and the *target object*; they may be the same in the case of a plain `JSError` or they may differ in the case of a `JSObject` that has `[[CapturedError]]` set.
The stack trace begins with a `toString()`ed version of the target object, continues with the formatted stack frames from the error object and is finally cached on the target object by redefining the `stack` as a regular property.

Reviewed By: kodafb

Differential Revision: D33906117

fbshipit-source-id: 059a7f935253ad54e8a1cb5e3810506252b0e33c