Construct runtime in forked process for death tests
Summary:
Death tests use `fork()` which does not copy over all the threads in
the parent process. This means that the Hades and Hermes runtime
threads do not exist in the forked process.

This can lead to deadlock, if for instance the Hades thread holds a
lock when `fork()` is called. It is likely the cause of test flakiness
in our TSAN tests.

There are two ways to fix this as far as I can tell, one is to use the
"threadsafe" death test style in GTest, which will restart the test
from the beginning in the new process.

This diff takes the second approach, to refactor our death tests
to always construct the runtime inside the forked process. This
allows for cleaner resource management.

Reviewed By: kodafb

Differential Revision: D34135340

fbshipit-source-id: 26ea4f975f73945ebbb65ca5027e27e66441343a