Use object ID for createPropNameIDFromString
Summary:
Our replay of `createPropNameIDFromString` isn't exactly accurate,
since we convert the `jsi::String` to a `std::string` and replay that
through the UTF-8 or ASCII codepaths instead. Change it to work like
any other function that deals with things in the JS heap.

This is also preparation for adding support for
`createPropNameIDFromSymbol`, which will need to working using an ID.

Note that this is fully backwards compatible with existing synth
traces, since the old format still works, it will just go through the
old codepath.

Reviewed By: kodafb

Differential Revision: D34034765

fbshipit-source-id: 9faaabb1d99751a5d57ea1d3f5148438e7586bcd