Use actual YG survival bytes in getDrainRate
Summary:
Instead of tracking the survival ratio in the YG and multiplying it by
the segment size, we can track the actual number of bytes that survive
a given collection. This will allow us to track external memory as part
of this metric later in the stack, since that cannot cleanly be
expressed as a survival ratio.

Note that this diff may have the net effect of slightly prolonging
incremental collections, because we were previously overestimating the
number of bytes that survive a YG collection in cases where the YG is
smaller than a segment.

Reviewed By: kodafb

Differential Revision: D28615404

fbshipit-source-id: 198f714a602d37afc75ef12607afb5100dc643ea