docs: add social banner in support of Ukraine (#688)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156668290-9c8ed034-471b-434f-af6e-c25aaf32d25c.png)

Pull Request resolved: https://github.com/facebook/hermes/pull/688

Reviewed By: neildhar

Differential Revision: D34649079

Pulled By: dmitryvinn-fb

fbshipit-source-id: f63d828a12c0480dfe32a12202e16276b217fe62