Implement ResolveLocale
Summary:
Implement a simple version of ResolveLocale which does not do any
validation against known locale data. We don't have direct access to
the locale data available on the platform, so this seems fine to skip
for now. We can revisit this by potentially encoding the allowed values
into this file, but I'd prefer to keep things simpler and more flexible
for now, and we can revisit it once the other parts of Intl are
implemented.

It is also built using std data structures to be consistent with the
other Intl code. The performance of this shouldn't be very important,
since it is only invoked when the constructor of things like Collator
is called, but we could move to a more efficient representation later,
once the other parts of Intl are implemented and we know how much
flexibility we really need.

Reviewed By: dmitryrykun

Differential Revision: D34886457

fbshipit-source-id: d6abb889e6520dbaec56623bc8759237fa1bd45d