Add utility functions (#652)
Summary:
This PR contains four utility functions - getOptionString, getOptionBool, getNumberOption and defaultNumberOption. They will be useful for implementing future Intl functions such as DateTimeFormat, NumberFormat, and Collator.

getOption has been split into getOptionString and getOptionBool, diverging from the ecma402 spec, which defines getOption as one function. This has been done to improve readability throughout the codebase, and it means that getOptionBool won’t have to pass through empty vectors.

getOptionString reviews a property, such as hour, and sees if there is a match with the valid properties pertaining to hour (numeric, 2-digit). If the property is empty, a fallback option will be returned. If there is a value, getOption will see if there is a match within the values vector, in which case the matching value will be returned, else if there is not a match, a RangeError will be raised. Likewise getOptionBool returns a bool value or empty fallback, but without the need for a vector.

getOptionNumber, instead of checking a string value match, checks whether the number option falls within an accepted range of minimum and maximum before returning through the defaultNumberFunction.

Pull Request resolved: https://github.com/facebook/hermes/pull/652

Test Plan: These four functions have been used within Intl.DateTimeFormat() successfully to validate the options that the user has given. They have passed test262 tests whereby the spec defines that a RangeError should be raised. Error messages display which option has caused an exception where possible. A RangeError will be thrown if a number is passed as an option property in getOptionString(), and vice versa if a string is passed through getOptionNumber().

Reviewed By: dmitryrykun

Differential Revision: D34364935

Pulled By: neildhar

fbshipit-source-id: 07bd3b3b77809f90964a34ddcfc182f4958bb760