Intl.Collator for Apple platforms
Summary:
Implement Intl.Collator on top of NSLocale. There isn't a perfect 1:1
mapping of Collator's options and NSLocale, so this uses a few
workarounds:
1. ignorePunctuation is implemented by stripping punctuation from the
string before doing the comparison, since there's no way to directly
ignore punctuation.
2. Sensitivity is implemented by setting specific flags on the
comparison to simulate the desired effect.
3. There is no way to directly specify "usage", so we instead supply
"search" as a collation in the event that no other collation is
specified.

It also does not check the validity of flags passed in, since we don't
have direct access to the underlying LocaleData. This is something we
can implement later, by enumerating the allowed flag values in this
file and checking against them.

Reviewed By: dmitryrykun

Differential Revision: D34675880

fbshipit-source-id: 19a8807e89116bd0b8a40a9f9ead3e61492b5448