Fix exception message from Error.prototype.toString
Summary:
Error.prototype.toString currently incorrectly uses the value after the
dyncast when reporting that the input type is incompatible. This is
always object, so the error incorrectly states that the receiver is an
object.

I also noticed that we don't currently handle Symbol in this switch,
instead just reporting it as "Value". This diff also adds correct
handling for that, which brings the errors in line with Node.

Reviewed By: jpporto

Differential Revision: D34969949

fbshipit-source-id: 25a55d3f62a2df206c5e8787beaa1354b3da7457